cloudrun-fetch-data/requirements.txt
flask==2.3.*
gunicorn==21.*
google-cloud-storage==2.*
ccxt==4.5.*
pandas==1.5.*
numpy==1.24.*